Add validateRegister method to authController
The authController was created to handle the registration and
authentication of users in the database. The validateRegister method
takes in the submission from the register form, sanitizes the fields,
and checks that they are valid. If all of the validations pass, it
passes it onto the next method. If any of them fail, it will return an
object with the errors.

The password-confrim logic is currently commented out until this is
added to the front-end form.