Update the validateRegister function
The validate register function was not formatted properly. In order to
make it work properly, it was broken out into two separate exports:
'validateRegister' and 'checkValidations.' The validateRegister runs
the validation tests against the form registration. After this is
completed the checkValidations function will be called which will see
if there are any errors. If there are errors, they will be sent to the
user. If there are not, the chain will continue.