Add authentication handling to login/register
The Login and Register components were updated to handle the logic
returned by the backend for user authentication. This includes
receiving a JSON response with a message, success or error including
information, and the url to redirect to. Upon successful registration
and/or login the user is redirected to the home page. Upon a failed
attempt, they remain on the page that they submitted the form.

TODO: Message display and handling needs to be added to the project
using the information in the JSON response.