Refactor Header to dynamically change links
The Header componentn now pulls the auth state from the Redux store
and dynamically changes the links in the header depending on the login
state. While the request is pending no links will be shown. When the
api request returns it will display either the logout links or the
register/login links.