Add action and reducer for tracking authentication
The action performs a request to the backend to determine if there is
a logged in user or not. It then stores this information in the Redux
store under the auth property. This can then be used on the front end
for selectively displaying information based on if the user is logged
in or not.