Add handling of removing a user favorite
If the return from the api yields an index, the entry was previously
in the database. It was removed on the back end, but needs to be
removed from the Redux store. The REMOVE_FAVORITE action type is
called which will slice the specified index out of the correct array.