Add function to poll Yelp API
The getFavoritesData method was created to fetch data for both a query
and the entire favorites object on the User model. If a location
property is included in the request the method will query the single
business. If now location property is passed, it will map over the
entire favorites object in the current user model and return the
populated data.