Make Passport logIn function use async/await
The login function had to be moved to use async await to resolve a
race condition that would manifest itself if trying to access req.user
immediately after login but before the information could be saved to
the session. To resolve this issue, req.session.save is called before
handling the redirects and uses the await keyword to ensure that the
data is stored in the session.