Refactor updateFavorites action and favoritesReducer
The updateFavorites action was modified to handle an additional API
call when there is not a location already stored in the user database.
In order to properly populate the Redux store with the added location,
the action must retrieve the Yelp data for that location. This API call
happens whenever a favorite is being added and passes the Yelp data as
the action.location property.

The dispatch actions were cleaned up for modifying favorites and only
pass the parameters that they need to run each respective action type.
Additionally, the favoritesReducer was updated to accept the Yelp
location data.