Connect DestinationDetails Component to Redux store
The component now receives its data from the destination state stored
in Redux. Upon the component mounting it calls the fetchDestination
action with the slug passed in as a route param. Due to a slight delay
in this data being populated, a renderContent function was created
which inserts in all of the data in the component. When the destination
props is null the page will render blank. As soon as the data is
populated, the entire page is rendered out.