Add destinations reference to User model
Another array was added to the User model under the favorites object.
This is an array of objectIds referring to the appropriate record of
each destination stored under the Destination model. In order to make
the queries more efficient, these fields are populated when findOne is
called on the User model. This will return the data of all the favorite
locations.

Currently, only the name, slug, and image properties are being passed
along as this is being used for the dashboard component. If additional
data is required, this projection can be updated.