Refactor HotelsList and Snackbar to be reusable
These components take in specific props for which yelp category and
locations it is updating. This will allow the components to be reusable
in the future for other categories besides hotels. These props also
affect which array in the database is updated. The props are required
for each instance of the components.