Update fetchDestinationCategory process
Previously each instance of HotelsList was calling the action to get
it's details from yelp for the category when the component mounted. The
tab switcher would cause a delay in this data being retrieved as the
component would not mount until it was the active tab. There was also
unnecessary API calls from the components remounting, but the data
already was in the Redux store. This was resolve by fetching this data
in the detailsTab component.