Connect Snackbar to auth state
The auth state was imported to get the raw array of yelpIds stored in
the database. The heart render and message checks the current yelpId
and sees if it is in the auth favorites array. This also allowed the
render to be more efficient. Additionally, the component was made
reusable depending on the yelpCategory prop passed to it.