Update logic for destinations
Some of the logic had to be updated in order to properly add and remove
destinations. The populate destinations field was moved from the User
model to the getFavoritesData method. Additionally, the use of includes
on updateFavorites did not work with the objectId. The code was
refactored to use indexOf which resolves the issue and maintains
functionality for yelpId strings.