Implement Circle Loader in Destinations page
Several more action types were added to Redux. The existing
FETCH_DESTINATION action was renamed to RECEIVE_DESTINATION. When the
page first loads the new FETCH_DESTINATION changes the isFetching flag
to true and displays the loader. When the information is received, the
RECEIVE_DESTINATION type saves the data and changes isFetching back to
false.

Additionally a new action and type was created to reset the destination
details to it's initial state. This action is called when the component
unmounts.