Add isFetching to flight workflow
* The destinationDetailsReducer now stores the flight data and an
isFetching flag.
* An additional type was created to RECEIVE_FLIGHTS. This turns off the
isFetching flag and replaces the contents of the array with the current
API response.
* The ticket.pop() and componentDidMount were removed from FlightsList.
These directly modified props which is an anti-pattern. This was
resolved through the clearing of destinationDetails state when the
component unmounts. Addiitonally, ticket.pop was not needed as the
contents of the flights.info array are replaced on each inquiry.
* Remove updateMessages actions from flightsList
* Add conditional rendering of loader for FlightsList component.
* Add renderFlightCost method to inject CircleLoader or cost data when
the user submits a request.